+++ tests/testit/test-plot.R
@@ -103,4 +103,4 @@ if (requireNamespace('tikzDevice', quietly = TRUE) &&
 }
 
 # https://github.com/yihui/knitr/issues/1166
-knit(text = "\\Sexpr{include_graphics('myfigure.pdf')}", quiet = TRUE)
+knit(text = "\\Sexpr{include_graphics('myfigure.pdf', error = FALSE)}", quiet = TRUE)